Closes #179

+++ src/main/java/fr/famivac/gestionnaire/web/utils/MontantConverter.java
@@ -1,27 +1,26 @@
 package fr.famivac.gestionnaire.web.utils;
 
+import jakarta.faces.component.UIComponent;
+import jakarta.faces.context.FacesContext;
+import jakarta.faces.convert.Converter;
+import jakarta.faces.convert.FacesConverter;
 import java.math.BigDecimal;
-import javax.faces.component.UIComponent;
-import javax.faces.context.FacesContext;
-import javax.faces.convert.Converter;
-import javax.faces.convert.FacesConverter;
 
 /**
- *
  * @author paoesco
  */
 @FacesConverter
 public class MontantConverter implements Converter<BigDecimal> {
 
-    @Override
-    public BigDecimal getAsObject(FacesContext context, UIComponent component, String value) {
-        return new BigDecimal(value);
-    }
+  @Override
+  public BigDecimal getAsObject(FacesContext context, UIComponent component, String value) {
+    return new BigDecimal(value);
+  }
 
-    @Override
-    public String getAsString(FacesContext context, UIComponent component, BigDecimal value) {
-        BigDecimal montant = (BigDecimal) value;
-        return montant.toPlainString();
-    }
+  @Override
+  public String getAsString(FacesContext context, UIComponent component, BigDecimal value) {
+    BigDecimal montant = value;
+    return montant.toPlainString();
+  }
 
 }
